Added documentation to composer interface and moved the unit test cases under gateway folder
